Accessing and modifying information in PCAGenomicSignatures

A set of accessor and setter generic functions to extract either the assay, colData, or metadata slots of a PCAGenomicSignatures object

Arguments

x
A PCAGenomicSignatures object

value
See details.

Setters

Setter method values (i.e., function(x) <- value):

geneSets<- : A character vector containing the name of gene sets used to annotate average loadings

studies<- : A list of character vectors containing gene sets used to annotate average loadings

gsea<- : A list of gseaResult objects.

metadata<- : A list object of metadata

`$<-` : A vector to replace the indicated column in colData

Accessors

All the accessors inherited from SummarizedExperiment are available and the additional accessors for PCAGenomicSignatures specific data are listed below.

model : Equivalent to the assays(x)$model accessor for convenience

geneSets : Access the metadata(x)$geneSets slot

studies : Access the colData(x)$studies slot

gsea : Access the colData(x)$gsea

`$` : Access a column in colData

trainingData : Access the trainingData slot

mesh : Access the trainingData(x)$MeSH slot

PCAsummary : Access the trainingData(x)$PCAsummary slot