Annotate top PCs from the dataset

This function finds the RAV scored highest with the top PCs of the dataset, including RAVs with the negative average silhouette.

annotatePC(
  PCnum,
  val_all,
  PCAmodel,
  n = 5,
  scoreCutoff = 0.5,
  nesCutoff = NULL,
  simplify = TRUE,
  abs = FALSE
)

Arguments

PCnum
PC number of your dataset you want to get the annotation results. It should be an integer value between 1 and 8.

val_all
The output from validate

PCAmodel
The PCAmodel used to generate the input for the argument, val_all.

n
An integer. Default is 5. The number of the top enriched pathaways to print out. If there are fewer than n pathways passed the cutoff, it will print out NA.

scoreCutoff
A numeric value for the minimum correlation. Default is 0.5.

nesCutoff
A numeric value for the minimum NES. Default is NULL and the suggested value is 3.

simplify
A logical. Under default (TRUE), the output will be a data frame with the number of column same as the length of PCnum argument, and the number of row same as the n argument. If it is set to FALSE, the output will be a list with the length of PCnum argument, where each element is a data frame containing detailed GSEA output of enriched pathways.

abs
Default is FALSE. If it's set to TRUE, the enriched pathways will be listed based on abs(NES).

Value

A data frame of a list based on the simplify argument. Check the output detail above.

Examples

data(miniPCAmodel)
library(bcellViper)
data(bcellViper)
val_all <- validate(dset, miniPCAmodel)
annotatePC(2, val_all, miniPCAmodel)

#>                                        PC2-RAV1076
#> 1                  REACTOME_METABOLISM_OF_PROTEINS
#> 2                                 KEGG_SPLICEOSOME
#> 3 REACTOME_ASSEMBLY_OF_THE_PRE_REPLICATIVE_COMPLEX
#> 4            REACTOME_LATE_PHASE_OF_HIV_LIFE_CYCLE
#> 5         MIPS_39S_RIBOSOMAL_SUBUNIT_MITOCHONDRIAL