Barplot GSEA output

Barplot GSEA output

msigdb_gsea(ind, PCAmodel, category = "C2", n = 10, pvalueCutoff = 0.5)

Arguments

ind
An interger. Index of avgLoading/PCcluster to apply GSEA.

PCAmodel
PCAGenomicSignature object.

category
A character vector representing MSigDB category. Options are "H","C1","C2"(default), "C3","C4","C5","C6", and "C7"

n
An interger. The number of top and bottom enriched pathways to plot. Default is 10.

pvalueCutoff
Cutoff for both pvalue and p.adjust. Default is 0.5.

Value

Barplot of GSEA output. Top and bottom n genesets based on NES are plotted and qvalues are denoted by color.