Two-dimentional PCA plot with the PC annotation

Two-dimentional PCA plot with the PC annotation

plotAnnotatedPCA(
  dataset,
  PCAmodel,
  PCs,
  val_all = NULL,
  scoreCutoff = 0.5,
  nesCutoff = NULL,
  color_by = NULL,
  color_lab = NULL,
  trimed_pathway_len = 45
)

Arguments

dataset
An expression matrix with genes (rows) x samples (columns)

PCAmodel
PCAGenomicSignatures-class object

PCs
A numeric vector length of 2. It should be between 1 and 8.

val_all
The output from validate

scoreCutoff
A numeric value for the minimum correlation. Default is 0.5.

nesCutoff
A numeric value for the minimum NES. Default is NULL and the suggested value is 3.

color_by
A named vector with the feature you want to color by. Name should be match with the sample names of the dataset.

color_lab
A name for color legend. If this argument is not provided, the color legend will be labeld as "Color By" by default.

trimed_pathway_len
Positive inter values, which is the display width of pathway names. Default is 45.

Value

Scatter plot and the table with annotation. If enriched pathway didn't pass the scoreCutoff the table will be labeled as "No significant pathways". If any enriched pathway didn't pass the nesCutoff, it will labeled as NA.