Validation result in data frame

Validation result in data frame

validatedSignatures(
  val_all,
  num.out = 5,
  scoreCutoff = NULL,
  swCutoff = NULL,
  clsizeCutoff = NULL,
  indexOnly = FALSE,
  whichPC = NULL
)

Arguments

val_all
An output matrix from validate function.

num.out
A number of highly validated PCclusters to output. Default is 5. If any of the cutoff parameters are provided, num.out or the number of filtered PCclusters, whichever smaller, will be chosen.

scoreCutoff
A numeric value for the minimum correlation.

swCutoff
A numeric value for the minimum average silhouette width.

clsizeCutoff
An integer value for the minimum cluster size.

indexOnly
A logical. Under the default (= FALSE), the detailed information on validated PCclusters, such as score, average silhouette width, cluster size, is printed. If it is set TRUE, only the PCcluster number will be printed.

whichPC
An integer value between 1 and 8. PC number of your data to check the validated signatures with. Under the default (NULL), it outputs top scored signatures with any PC of your data.

Value

A subset of the input matrix, which meets the given condition.

Examples

data(miniPCAmodel)
library(bcellViper)
data(bcellViper)
val_all <- validate(dset, miniPCAmodel)
validatedSignatures(val_all, num.out = 3, scoreCutoff = 0)

#>                   score PC          sw cl_size cl_num
#> PCcluster1076 0.5950767  2 -0.04447124      10   1076
#> PCcluster2538 0.5838616  2  0.06996166       4   2538
#> PCcluster338  0.5709072  2 -0.04683319      21    338

#                   score PC          sw cl_size cl_num
# PCcluster1076 0.5950767  2 -0.04447124      10      1
# PCcluster2538 0.5838616  2  0.06996166       4      2
# PCcluster338  0.5709072  2 -0.04683319      21      3