Validate new datasets

Validate new datasets

validate(dataset, PCAmodel, method = "pearson", maxFrom = "PC", level = "max")

Arguments

dataset
Single or a list of SummarizedExperiment (RangedSummarizedExperiment, ExpressionSet or matrix) object(s). Gene names should be in 'symbol' format.

PCAmodel
PCAGenomicSignature object. You can also provide signature model matrix directly.

method
A character string indicating which correlation coefficient is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.

maxFrom
Select whether to display the maximum value from dataset's PCs or avgLoadings. Under the default (maxFrom="PC"), the maximum correlation coefficient from top 8 PCs for each avgLoading will be selected as an output.If you choose (maxFrom="avgLoading"), the avgLoading with the maximum correlation coefficient with each PC will be in the output.

level
Defibe how to ouput validation in two different forms, c("max", "all"). Default is "max", which outputs the matrix containing only the maximum coefficient. To get the coefficient of all 8 PCs, set this argument as "all".

Value

A matrix containing the maximum pearson correlation coefficient between the top 8 PCs of the dataset(s) and pre-calculated average loadings of training datasets. Each row represents a new dataset for test, and each column represents clusters from training datasets. If level = "all", a list containing the matrices of the pearson correlation coefficient between all top 8 PCs of the dataset(s) and avgLoading.